задание 9 -исправления
