+++ js/permanent.js
@@ -1,0 +1,18 @@
+const Screens = {
+  INTRO: `intro`,
+  GREETING: `greeting`,
+  RULES: `rules`,
+  GAME_1: `game-1`,
+  GAME_2: `game-2`,
+  GAME_3: `game-3`,
+  STAT: `stat`
+};
+
+const Velocities = {
+  NORMAL_MODE: `correct`,
+  SLOW_MODE: `slow`,
+  FAST_MODE: `fast`,
+  WRONG_MODE: `wrong`
+};
+
+export {Screens, Velocities};
